github: Add information about the recheck command
We've seen that people sometimes use a wrong command (eg /recheck) when
they try to rerun jobs. This adds the guidance about the correct
command to avoid such wrong usage.

Also this change removes the reference to the upstream infra guide,
because most of the items described in the doc (eg. gerrit) is not
really relevant to the projects hosted in github.

Change-Id: I5e5a244d81f68c159e7752c6aaf192e33481b514