Match any line that contains keywords
Current jobs are not triggering if someone add
additional info after recheck, ex[1]

$: This symbol indicates the end of the string.
It ensures that the regex pattern matches only if
the specified words occur at the end of the input string.

Getting rid of \s*$ to match any line that contais these keywors
and not literally checking for the exact word.

[1] https://github.com/openstack-k8s-operators/openstack-ansibleee-operator/pull/190#issuecomment-1647608530

Change-Id: Id0529ce6a1426cfad9fcd383a9a31d2dcf728b68