Use "localhost" instead of "controller"
We're facing an issue with the current way things are done: since we
inherit our jobs from base, we're also inheriting from their "post-run"
- and there, the ssh key are apparently cleaned. Since we're using this
playbook as a "cleanup" task, it comes after post-run - so we don't have
SSH access anymore.

Trying to switch to "localhost", hoping it will do the right thing and
avoid the ssh failure.

Change-Id: Ic10d162daa58923d26017d6d40ce8156771d6fde