Make openstack client pip package user wide
As now the bootstrap playbooks are run targeting localhost
it possible to not have enough permissions (or even not
having sudo) to install the pip package as a system package.

To ensure that the openstack client executable is properly
found now all calls to the client are made pointing to the
absolute path of the installed binary.

Change-Id: I38cc209c9b33fffefd8bce096690a7aba84daa48