Configurable cloud-name in boostrap-ci-network
As resources from nodepool might be from
different clouds/projects (depending on label/nodeset),
there is need for ability to select in which project
to bootstrap the network resources.

This makes it possible to specify which subkey
of cloud_secrets should be used,
as also uses that value as name of the cloud in
rendered temporary clouds.yaml file.

Change-Id: I0cd0a5e24a050b6d78372080a7df2d771ccdb87f