Allow skipping the cleanup if a file exists
We are about to add some logic to skip the cleanup of the bootstrapped
network if there is an authold set for a given file. That logic will be
stored on ci-framework and it will allow jobs running the multinode
layout to be holded without the need of adding explicetly the
crc_ci_bootstrap_skip_cleanup variable, that sometimes requires making a
new patch an loosing the environment. That's specially unfortunate when
looking for non consistent errors.

Change-Id: I2497cc1d947a9539fa8430cf8797e01471a4f968