Do not cap ruamel.yaml in buildsys-tags-update
We have to uncap ruamel.yaml to get latest one and be consistent
with other jobs not capping it.
Version 0.15.0 is available up to py3.6 [1], so as we'll have to
move to CS9 (py39) as nodeset in the near future, we should
switch now.

Note that ruamel.yaml, git-review and sh are installed with releng after.

[1] https://pypi.org/project/ruamel.yaml/0.15.0/

Change-Id: I1ef1e4c6f136beba42e564cd65ba41f1b59037fa