Remove subscription-manager unregister from cleanup-run plays
In [1] we added the unregister play as cleanup-run. We do not get logs
from cleanup-run but from watching console this play is not working.

For unknown reason zuul cannot reach the standalone host:

       ERROR  "msg": "Data could not be sent to remote host

As this play does not need secrets (unlike register) lets also move this
to rdo-jobs, in v9 of [2], also noting that child post-run are executed
before parent plays.

[1] https://review.rdoproject.org/r/c/config/+/50589
[2] https://review.rdoproject.org/r/c/rdo-jobs/+/50590

Change-Id: I4366416ab8e677357faffad916bf8caa3e359a8f