Include project-templates from projects
Allow project-templates defined in a project to be included.
This particularly applies to infrawatch, which has STF components, each
in their own repo, but which use the same set of jobs for testing.

Other groups of project may also want to define custom project templates
as well e.g. a team may want to have integration tests for a sub-set of
components.

Related: https://github.com/infrawatch/service-telemetry-operator/pull/514
Change-Id: Ia6284fa3de81a464440dc3583ad63e38b7243169