Add new pipeline openstack-periodic-daily2
In order to clean up resources which were used by the old tripleo
periodic pipelines and RDO was reusing for weirdo jobs, I am proposing
to create a new openstack-periodic-daily2 pipeline. I will move the
pipelines which are currently running in some of the
openstack-periodic-integration-stable pipelines to it.

I prefer not to run all the periodic jobs in the sam
openstack-periodic-daily to split the load among two different
schedules.

Change-Id: Ic3fa08b3a3341fa30e5fdc2b0cbdaf7de3baff5d