[crc] Wait for the cluster to come up if there's a custom bundle used
Using a non-default bundle in CRC, the time taken to start the cluster
can exceed the timeout in crc start.
This causes the crc start command to fail (as seen in [1]), but the
cluster still comes up afterwards.

One solution to this would be to increase the resources given to the CRC
VM, but this will cause all the jobs to use more resources.

If we ignore the errors from the ``crc start``, a new task can be added
to check the cluster status and wait for a reasonable amount of time
for the cluster to come up. This is tested in [2] by running the
playbooks in ``post-run`` instead of ``run``

[1] https://review.rdoproject.org/zuul/build/908bdc30ed104069b001d6ae73311251/log/controller/crc-start.log
[2] https://github.com/infrawatch/service-telemetry-operator/pull/523

Change-Id: Idc736f718b78dd8fc9a39fd2386b92a4def9eb23