Replace SA pull-secret with a pabrodri's personal one
Since we moved from not preserving the pull-secret to preserve
it for the entire run of the job we now want to avoid leaking
our service account pull-secret. To avoid leaking that one that
is associated to the organization we have created a regular
customer one using pabrodri's account.
All the ci-framework jobs that consumed the SA pull-secret
switches to the new one.

Change-Id: I4e911747ffdb6e9f627f673bdd3fed8c403f0831