Add condition to ensure that machineconfigpool state has changed
The machineconfigpool when it's degraded, we re-create the missing
manifest to satisfy the condition, so later the machineconfigpool
are set as non-degraded.
Normally, the update is simply and it is done in few secods, but
on some infras, where is high load, the update might take a while.

This commit is adding one more condition to ensure that the
machineconfigpool has been updated before continue.

Change-Id: I22de2aff1fb43fcf263fdfa9fb5cb5780b29c4fe