Add workaround for new CRC 2.34.1 to speedup marketplace convalescence
In the new CRC version (2.34.1, it is possible that before we inject
the pull-secret.txt into /var/lib/kubelet/config.json, the openshift-marketplace
operator would try to re-create pods (the openshift-marketplace operator
changed the imagePullPolicy to always), which can take long time
(over 20 minutes) to make the OpenShift cluster up and ready with all
deployed operators.
To avoid waiting for the pods, let's trigger them to remove and
after recreating, it would be working on first retry.

Change-Id: Ibff45547c419fe3333d4524b12108f37649ccd8b