Add pull-secret also to the Openshift config secret
We inject directly the pull-secret content into the
/var/lib/kubelet/config.json file, where later it will be manipulated
by the mcp (machineconfigpool) operator (machine-config-operator).
To avoid sutation, that we inject the pull-secret into the config.json
and the mcp will set `{}` to that file, let's set the
content also to the secret, so the rendered-master configuration
should contain the new file content.
More info about machineconfigpool work [1].

[1] https://www.redhat.com/en/blog/the-consequences-of-pausing-machineconfig-pools-in-openshifts-machine-config-operator-mco

Change-Id: Ibcdd1576678b7373a6707ed4bb4c359a4a744336