Add workaround for machineconfigpool when files are modified
The MCO (machine config operator) when its spotted, that file
on the host is different comparing to what it got in the manifest,
it raise an error and the machineconfigpool is marker ad DEGRADED.
We are doing some operations on the "sensitive" files, like:

    - /var/lib/kubelet/config.json
    - /etc/containers/registries.conf
    - /etc/containers/policy.json

And optionally:

    - /etc/node-sizing-enabled.env
    - /etc/node-sizing.env

but on making changes on that file, CRC host reboots and got
troubles to start the kubelet service (because they are created by
other script that is executed before kubelet systemd service starts).

The best way to not have mcp degraded is to remove the condition or
change the condition state to current one, but it requires
modification in the CRC code. We already remove the condition,
so the new rendered config should not include files that we change.
The issue with that way is, that the mcp can not be paused on doing
the `crc bundle generate` command, because if will fail on cleaning up
the pull-secret key, so to avoid changes in the CRC, we are doing
that after the CRC extracted VM start. It makes one more problem: the
MCO operator render the new rendered-master-* config, which is "watching"
on the file that we are modifying later, so it is required to dump
new rendered config manifest, remove condition (same as we are doing
in the missing manifest) and apply the new manifest, so the MCO
will not "watch" on the files, what we are modifying.

Change-Id: I98885af0868e2fa9940f6ce9fae62e7f4f6563d1
Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>