Fix base64 encode for pull-secret
Current Ansible version (2.13.7) that is running on the Zuul CI have troubles to
parse the pull-secret content and encode to the base64.
For example:

    openshift_pull_secret: {"test":"test"}
    openshift_pull_secret_alt: |
      {"test":"test"}

By calling variables:

    {{ openshift_pull_secret }}
    {{ openshift_pull_secret_alt }}

Will return result:

    {'test': 'test'}
    {"test":"test"}

Which is problematic for the b64encode filter, that later raises an
error:

    "stderr_lines": ["The Secret \"pull-secret\" is invalid:
    data[.dockerconfigjson]: Invalid value: \"<secret contents redacted>\": invalid character '\\''
    looking for beginning of object key string"], "stdout": "", "stdout_lines": []}

The current error is not raising in newer Ansible version (tested on 2.16.5).

Change-Id: I2f4aa87b6d90599c87d1eac2db69e9525eb7d139
Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>