Add a workaround for many rendered-master machineconfigs
It happens, that the OpenShift starts longer as usual and the
MCO (machine config operator) creates additional render-master-* config,
which is not needed, but it has bigger prio. In that case, let's
remove conflicting keys from the mc (machineconfig) to make the
mcp work.
Also add a task to remove machineconfigpool dir on finish.

Change-Id: I09988c49bdc3054c27e2f911befadcb58c592409