Filter the CRC password from the whole file
When running CRC with OKD preset, the output of
crc-start.log is more verbose than the default preset.
As a consequence, by using tail we don't have the kubeadmin
password in the output. So we end up with 'developer' password for kubeadmin.
By cat'ing the whole file we get the the kubeadmin password.

Change-Id: I41b28d1995fb85776d22345898dee1e01b7d1165