Remove problematic keys also from 00-master and 00-worker machineconfig
The machineconfigpool is once again degraded after the NextGen Ansible tasks
are applied. Normally the machineconfigpools are set to pause, but it
might happen, that the CRC host is rebooted and the paused parameter
after is set to false, which might set the machineconfigpool to
degraded.
This commit is removing the problematic keys from all important
machineconfig that contains the files to track.
NOTE: some machineconfig rendered-{master,worker}-* might contain the
files to track, but it should not affect the mcp state.

Change-Id: Ib08767e610c934103c09e205bc3bbaad4a605874