Remove problematic keys from mc; add custom machineconfig for kubelet config
The machineconfigpool is once again degraded after the NextGen Ansible tasks
are applied. Normally the machineconfigpools are set to pause, but it
might happen, that the CRC host is rebooted and the paused parameter
after is set to false, which might set the machineconfigpool to
degraded.
This commit is removing the problematic keys from all important
machineconfig that contains the files to track and it creates own
MachineConfig manifest with the new /var/lib/kubelet/config.json file,
so the MachineConfigPool should not be degraded after restart.

IMPORTANT NOTE: Similar manifest(s) should be done for other "sensitive" files,
that we modify in the CI [1] if those files are changed to avoid
MCP degradation.

[1] https://review.rdoproject.org/r/plugins/gitiles/config/+/refs/heads/master/roles/prepare-crc-extracted/files/remove-machineconfigpool-keys.py#12

Change-Id: Ib08767e610c934103c09e205bc3bbaad4a605874