Update github-check URLs
When a job is triggered in Github, Zuul provides a URL to check the
status at various stages of the job lifecycle. Currently we use the
`review.rdoproject.org` hostname. This works but when
a user needs to authenticate they need to browse to `softwarefactory-project.io`. This patch changes the link that will be
provided to GitHub so everything is on the same domain.

Change-Id: I660fc811d2a07e31c0ac11442e2a25203b0c56f6