Add crc-cloud base job and role
The crc-cloud will help infra team to avoid future issues related to the
image, due the image is using crcbundle file, what is provided by the
CRC team [1]. That image would only include the service, that
will be responsible for injecting the cloud-init SSH keys - all the rest
is pure bundle image.

NOTE: This commit is using temporary crc-cloud repository [2] that contains
a fix for the pull-secret issue, which is missing to spawn the OpenShift
console or OpenShift marketplace.

[1] https://mirror.openshift.com/pub/openshift-v4/clients/crc/bundles/openshift/
[2] https://github.com/danpawlik/crc-cloud

Change-Id: I2cb4f5de2a75eaab3cd6d9413ddf9ba9b939d086