Disable machineconfigpool for crc node; remove reconfigure kubelet
There was many attempts to keep the machineconfigpool not degraded, but
in many CI jobs, some files on the system are modified. Therefore
later it is degraded. Disabling the mcp should help some teams that
are relying on the mcp degradation state.
This commit also remove the reconfigure kubelet service, because we
stopped using it long time ago.

Change-Id: Ic3dc4deb4d6e8918f8ba571b506adf45f36019db