Do not edit kubelet config.json; drop reconfigure kubelet
This change creates a new generic secret [1] that would be used later to
get access to limited repositories that are needed to deploy e.g.
OpenShift Marketplace operator. With that way, we don't need to change
anything in the /var/lib/kubelet/config.json file, so the
machineconfigpool should not raise an error, that are degraded.

[1] https://docs.openshift.com/container-platform/4.15/openshift_images/managing_images/using-image-pull-secrets.html#images-allow-pods-to-reference-images-from-secure-registries_using-image-pull-secrets

Change-Id: Ic3dc4deb4d6e8918f8ba571b506adf45f36019db