Configure dnsmasq and NetworkManager if crc-pre service not available
The kubelet should not be started before applying the changes in the
dnsmasq or NetworkManager. Normally, the crc-pre.service [1] [2] is doing
that, but it makes some complications on some infras to use that script,
so we disable it [3] and future CRC images will not have that service.
So far, we need to run those commands to be able to login to the
OpenShift cluster to verify if all pods are up and running.

[1] https://github.com/softwarefactory-project/sf-infra/blob/master/roles/next-gen/extracted-crc/files/crc-pre.service
[2] https://github.com/softwarefactory-project/sf-infra/blob/master/roles/next-gen/extracted-crc/files/configure-pre-crc.sh
[3] https://softwarefactory-project.io/r/c/software-factory/sf-infra/+/31954

Change-Id: I9fc57886db8b74e7e0e2fe7fe74b3bffd26362a1