Change machineconfig generated-registries priority
In the CRC, there are machineconfigs responsible to track files that
later are modified by the OpenStack K8S Operators projects like:
/etc/containers/policy.json or /etc/containers/registries.conf
with the higest priority (99).
Therefore, the machineconfigpool is later degraded.
To avoid situation, that we need to dump the machineconfig, then find in
the list a dictionary, that contains the file, we set the current
machineconfig to lower priority, so later we will be able in the CI
create another mc that will have prio 99 and that will be applied.
After that, the machineconfigpool should not be degraded.

Change-Id: I6b00bfd153cdc0e6bc78dc4b08fdd32e70f04e06