Change machineconfig generated-registries priority
The CRC is creating the machineconfig with the highest priority, which
is later complicating the CI workflow. The Openstack K8S projects are
modifying the /etc/containers/policy.json or
/etc/containers/registries.conf, which are included in the mentioned
machineconfig. Therefore, the machineconfigpool is later degraded.
To avoid situation, that we need to dump the machineconfig, then find in
the list a dictionary, that contains the file, we set the current
machineconfig to lower priority, so later we will be able in the CI
create another mc that will have prio 99 and that will be applied.
After that, the machineconfigpool should not be degraded.

Change-Id: I6b00bfd153cdc0e6bc78dc4b08fdd32e70f04e06