Pause the machineconfigpool changes to be applied
Currently our CI got many failures because of the openshift-marketplace
that can not be started properly. After checking, the config.json
in the /var/lib/kubelet dir is cleaned.
We already provide a machineconfig that creates the file with proper
content [1], but we would like to have an option just to pause the mcp
to apply changes.

[1] https://review.rdoproject.org/r/c/config/+/53842/

Change-Id: I1a4c31372f419cfc5ddc58fca538839ccb3f59d5