Simplify the machineconfigpool degradation
Temporary, we will pause waiting for the machineconfigpool to not be
degraded (we don't want to break the CI tests).
In the near future, when all "sensitive data" (files that are watched
by the machineconfig) will be not edited by the Ansible, we will
enable back fixing the machineconfigpool degradation.

Change-Id: Ib72346af4cec6b89739a26c9b49d1e7fc38eb853