Wait for config.json to be not empty; cleanup marketplace on the end
When the parameter wait_for_mcp is set to true, we are waiting for
the machineconfig to be not degraded, but during that time, host is
still in the "Updating" state. That takes few minutes, but we can
continue if the file /var/lib/kubelet/config.json is not empty.
Also moved the workaround for flapping OpenShift Marketplace pods
to the end of the base job, because after the machineconfig daemon
trigger the force reconfiguration, host is rebooted. After comming back,
the marketplace can not be spawned. It takes a long time to estabilish
the operator back.

Change-Id: Id0d049af89c6bb4245b1375a13dc280038a4c396