Add udev rules for main CRC interface name
When the Machineconfig daemon triggers reboot on the CRC host,
it happens that the interface name has been changed from enp6s0
to enp7s0. That raises many complications in the CI tests, where
scripts and other Ansible tasks are relaying on that interface name.
By adding the udev rule, the NetworkManager should keep the interface
name as it was earlier.

Change-Id: I6066a0ee5343bb88910aace54f61d655987505e2