Re-run clustersetup.sh script and remove machineconfig pods
From time to timem even retry does not help to deploy the CRC by using
the clustersetup.sh script. What is mentioned in the logs:

    [INF] waiting for machineconfig to become available try 18, hang on....
    [INF] waiting for machineconfig to become available try 19, hang on....
    [ERR] impossible to get resource machineconfig

The daemon pod is in crashloopbackoff state. Let's remove all pods
from that namespace and let's wait for the recreation of it (similar
workaround was done for CRC and OpenShift Marketplace).

Change-Id: I924e3535310b9a6fabe3ddecbc6051603dc5a32a