Move marketplace recreation earlier; check if oc returns value
It happens, that the OpenShift API is not stable enough, so the oc
command returns:

    The connection to the server api.crc.testing:6443 was refused - did you specify the right host or port?

which is wrong, but after few seconds, it is stable.
That patch will ensure that the API returns something, not just an
error.
Also moved the part for recreating the OpenShift Marketplace to be
executed earlier - it is required for older CRC images version.

Change-Id: I03895278db21cb98e0c8bcdae99bc4a963de708e