Introduce trigger_branch_name var in trigger job
Trigger job works on file changes. Based on trigger_job_name
var, it creates the file on which required jobs runs downstream.

Since we are planning to run trigger jobs on different branches
i.e. 18-0-0-proposed. So we need to make trigger file based on
branch name and job name.

It will help to define jobs on jobs_name and branch_name.

Change-Id: Idbddb1d5cc2d464125a1ee606352b33b6f84deba
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>