Make wait for the OpenShift API after clustersetup.sh script
From time to time, no matter if we are waiting for the
MCP (machineconfigpool) or not, the OpenShift API crash and the
pods are recreated.
Let's set temporary 2 minutes value to wait for the cluster to be
stable. From my local tests, 1 minute should be enough.

NOTE: It is difficult to make a check task to wait for all pods to be
available in some namespace, due the "crash" might happen in
few namespaces.

Change-Id: I892a17ac41d4ac640d13fe29e6a0290b38ce3c43