Remove creating additional domains for crc-cloud; improve /etc/resolv.conf
We don't want to add additional domains to the CRC cluster.
That part of the script is not needed to make the cluster stable.
Also improve resolv.conf file and add a task to login to the OpenShift
just to be sure that we can remove openshift-marketplace containers
before continue.

Change-Id: Id8d8ce6006838866fba270d97da4bbe8c1095dee