Restart kubelet on CRC host if login fails after long time
It might be a situation that the services to start properly require
more time as expected. Normally we could reboot the host, but there
is a chance, that the etcd would be on ramdisk. We can add condition,
that if etcd on ramdisk, we can reboot host, otherwise restart kubelet,
but it does not make sense to "rescue" the CI job if it makes troubles
from the beginning.

Change-Id: If392229815827ecced7b92daeb6d720f29e781fb