Add etcd slower profile option in CRC; set default CRC version to 2.48
The etcd service is crashing from time to time during the
CI tests on high loaded hypervisors. It is because etcd is very
sensitive for the disk delays, which are high in our CI - even
when the quota is set.

This commit is setting etcd slower profile by default, that might
help our CI jobs to be in better shape.

Change-Id: Ia19b327731b90639eea1a950e72185fc4dfdae68