Add IOPS and read/write limitation to nested CRC
To avoid issue with the noisy neighbour that is raising issue
with the etcd, which is killing the OpenShift API, we need to
make any workaround to optimize better our CI jobs.
This commit is proposing IOPS and Read/Write limitation
to the nested CRC, so even when the VMs would be on same host,
there should not be any "spikes" in IOPS usage.

Change-Id: I3ccaefba00416ce5d0ccd1ca091b72216c237779
Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>