Add generic crc job for crc-cloud and crc-extracted
This change brings just a new layer that will be dropped after
migration from crc-extracted to crc-cloud is done.
To avoid unexpected complication and many pull requests
that will be failing because of missmatch for nodesets
in parents and in child (e.g. parent would be crc-cloud where
child is overwriting nodeset to crc-extracted, so all the
other (childs) nodesets needs to be updated in the same time),
we create such base "layer".

Co-Authored-By: Joan Francesc Gilabert <jgilaber@redhat.com>
Co-Authored-By: Bhagyashri Shewale <bshewale@redhat.com>

Change-Id: I128d38f47ed69e8621b7934bc55aa896c494d70a
Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>