Use a fact for application_credential
Use set_fact task to define application_credential variable
in the trusted context before it's needed.

Removes the vars from tasks that used that.

When defining variables with `vars:` in a task, Ansible/
Zuul change the evaluation context. The secret variables
which are `!trusted` become inaccessible during template
rendering.

By using `set_fact` the application_credential variable is
defined in the trusted context and can then be used in
combination/togheter with the truested secret variables
in subsequent tasks.

Change-Id: I52df72e2621b7ea89a1c9490abcde24fa9052877