Change image build *_dir facts
Using `zuul.executor` src/work_root is a missunderstanding,
wrong context - it must use the path on the node i.e ansible_user_dir.

The `prepare-workspace` will ensure the same structure is used when
ensuring src repos are created on the node so the
zuul.project['...'].src_dir is correct.

Change-Id: I57ab90f81e24c52db81c6e5d4a1b7529f96de00d