add shadow directive to openstack/devstack
This change uses the shadow directive to declare
that if there are job conflict between the josb defiend
in upstream openstack/devstack and rdo config then
the rdo jobs shoudl win and the devstack ones should
be ignored.

Without this change its a error for the job to exist
in both places. This causes issues with testing
the supprot for Centos 10 stream in devstack as
it is adding an upstream version of the existing
rdo devstack-platform-centos-10-stream job.

Needed-By: https://review.opendev.org/c/openstack/devstack/+/937251
Signed-off-by: Sean Mooney <work@seanmooney.info>
Change-Id: Ie251b8f91964b4da26eacf82b0fd6eda887cb459