Update cores for compute projects
Kashyap, Sean, and Lee have not been part of the Compute team for a
while now. Update the core groups for the various Compute projects to
Sylvain and Balazs. This continues the work from
https://review.rdoproject.org/r/c/rdoinfo/+/58105 and
https://review.rdoproject.org/r/c/config/+/57784. puppet-placement is
potentially controversial because no one currently on the Compute team
has much context for it, but then again it has literally never had any
patches proposed (at the time of this writing), so the whole thing is
moot.

Change-Id: I6b6df313e9f2cf725a8c9a7724783c6b70a16062