Re-factor hotstack plays
This change removed the "run" playbook, instead it run's
two `pre-run` playbooks.

pre_stage1:
 Bootstaps the application credentials and hotstack cloud secret
pre_stage2:
 Set up the infra using heat stack
 Bootstrap the hotstack "controller / infra" instance
 Install OCP
 Deploy RedFish virtual BMC on the OCP cluster
 Remove cloud credentials

Also in the post stage, re-write the clouds.yaml file in both user
and application credential cases, so that the resources (heat
stack) can be deleted. (Previously not required because the
credential file was not deleted at the "end" of pre-run plays)

The goal with this change, is to run the RHOSO deployment and any
test plays from a `run` playbook in non-config repo.

Change-Id: I9846b4f019b9802eb332f6989579f61ad0371e1b