[nested crc] Ensure all cluster operators are available before continue
After making various changes on the CRC node like restart cri-o,
set etcd slow profile, set disk limitation, we verify the
cluster status if it is stable. In some corner cases, we see
that the cluster is not stable enough, so commands that goes to
openshift apiserver are failing.
Wait for cluster operators to be stable before continuation.

Change-Id: Ideb064c3ff30f2f7df2946e11bbfd5b98a60b7a1
Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>