[crc-cloud] Create cluster-cert-approver workaround
The new CRC images based on OCP 4.20.1 have expired certificate,
so the cluster can not by fully deployed properly.
The same issue is in newer CRC version 2.57 (based on OCP 4.20.5).
Until the CRC team does not fix it, we need to use the workaround
to make the CI jobs stable.

Change-Id: I65937abf62968c9803208978c3420d076d4d637a
Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>